backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
openai-whisper==20231117
TTS==0.20.6
pydub==0.25.1
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4z